general scatter

+++ lang/math.h
@@ -35,7 +35,7 @@ struct Matrix {
 
   Matrix map(const std::function<Expr(const Expr &)> &f) const {
     Matrix ret(n, m);
-    for (int i = 0; i < entries.size(); i++) {
+    for (int i = 0; i < (int)entries.size(); i++) {
       ret.entries[i] = f(entries[i]);
     }
     return ret;
